Rename: pages폴더 삭제, app으로 이동
